+++ src/is_nothing.ts
@@ -1,6 +1,6 @@
+import { Nothing } from './utilities';
+
 /** Checks if value is `null` or `undefined`. */
-export function isNothing<T>(
-  value: T | null | undefined
-): value is undefined | null {
+export function isNothing<T>(value: T | Nothing): value is Nothing {
   return value === null || value === undefined;
 }